# microservicios/perfiles-service/internal/httpapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API de Gestión de Perfiles",
    "version": "1.0.0",
    "description": "Perfiles creados y sincronizados de forma asíncrona a partir de eventos de empleados. Los campos personales editables se modifican por REST."
  },
  "servers": [
    { "url": "/" }
  ],
  "paths": {
    "/perfiles": {
      "get": {
        "summary": "Listar todos los perfiles",
        "description": "Retorna perfiles activos y archivados para conservar trazabilidad.",
        "tags": ["Perfiles"],
        "responses": {
          "200": {
            "description": "Listado de perfiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Perfil" }
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/perfiles/{empleadoId}": {
      "get": {
        "summary": "Consultar perfil por empleado",
        "tags": ["Perfiles"],
        "parameters": [
          {
            "name": "empleadoId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "E001"
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Perfil" }
              }
            }
          },
          "404": {
            "description": "Perfil no existe",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Actualizar datos editables del perfil",
        "description": "Permite modificar teléfono, dirección, ciudad y biografía. Nombre y email se sincronizan desde empleado.actualizado.",
        "tags": ["Perfiles"],
        "parameters": [
          {
            "name": "empleadoId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "E001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ActualizarPerfil" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil actualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Perfil" }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Perfil no existe",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "409": {
            "description": "Perfil archivado; no admite modificaciones",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Perfil": {
        "type": "object",
        "required": [
          "id",
          "empleadoId",
          "nombre",
          "email",
          "telefono",
          "direccion",
          "ciudad",
          "biografia",
          "fechaCreacion",
          "archivado"
        ],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "empleadoId": { "type": "string", "example": "E001" },
          "nombre": { "type": "string", "example": "Juan" },
          "email": { "type": "string", "format": "email", "example": "juan@empresa.com" },
          "telefono": { "type": "string", "example": "3001234567" },
          "direccion": { "type": "string", "example": "Carrera 15 # 12-34" },
          "ciudad": { "type": "string", "example": "Armenia" },
          "biografia": { "type": "string", "example": "Ingeniero de sistemas" },
          "fechaCreacion": { "type": "string", "format": "date-time" },
          "archivado": { "type": "boolean", "example": false },
          "fechaArchivado": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ActualizarPerfil": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "telefono": { "type": "string", "example": "3001234567" },
          "direccion": { "type": "string", "example": "Carrera 15 # 12-34" },
          "ciudad": { "type": "string", "example": "Armenia" },
          "biografia": { "type": "string", "example": "Ingeniero de sistemas" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "boolean", "example": true },
          "mensaje": { "type": "string" }
        }
      }
    }
  }
}